feat(search): consolidate Cmd+Shift+F into Cmd+K command palette (#134)
Merge GlobalSearch (file content + terminal scrollback + open panel search)
into the command palette so there is one unified search interface.

- Replace git-ls-files name-only search with fsSearch (name + content)
- Add terminal scrollback and open panel matching to Cmd+K results
- Remove GlobalSearch component, shortcut binding, and uiStore state
- Redirect native menu "Find in Files" to command palette
