+++ src/renderer/hooks/useShortcuts.ts
@@ -136,7 +136,7 @@ export function useShortcuts(): void {
           canvasStore().autoLayout()
           break
         case 'globalSearch':
-          useUIStore.getState().setShowGlobalSearch(true)
+          useUIStore.getState().setShowCommandPalette(true)
           break
         case 'undo':
           canvasStore().undo()
@@ -223,7 +223,7 @@ export function useShortcuts(): void {
       if (e.key === 'Escape') {
         if (terminalHasFocus) return
         const ui = useUIStore.getState()
-        if (!ui.showCommandPalette && !ui.showNodeSwitcher && !ui.showGlobalSearch) {
+        if (!ui.showCommandPalette && !ui.showNodeSwitcher) {
           canvasStore().clearSelection()
           // Don't prevent default — Escape might also close other things
           return
